fix: split chat detail into server page + client component for static export compat
